+++ module_test/test_load_encoded_id.cpp
@@ -2,7 +2,7 @@
 
 int main() {
     std::string file_name = "encoded_prompt.bin";
-    std::vector<float> encoded_id = get_encoded_id(file_name);
+    std::vector<size_t> encoded_id = get_encoded_id(file_name);
     // inference();
     return 0;
 }
